Completes text on first click during dialogue phase
Typewriter component exposes if it's typing, and keeps track
of what it's writing and to where so it can stop typing as well.
When it stops typing it completes the text, printing out the entire
thing, and stopping the animation.

DialogueManager will handle if the Typewriting component should
finish typing or just move on, based on if it is currently typing.